refactor(refs): co-locate resolve-ref with existing validation utilities
Addresses PR #683 review feedback:

- Move `src/utils/refs/resolve-ref.ts` → `src/utils/validation/resolve-ref.ts`
  (co-located with `name-validator.ts` — both are ValidationError-throwing
  input normalizers). Removes the new `utils/refs/` folder Raina flagged;
  the test file follows into `tests/unit/utils/validation/`.
- Switch the caller-label error test from a `try/catch` block to the
  `.rejects.toBeInstanceOf` + `.catch()` pattern that every sibling test
  in the file uses — matches the convention rule about consistent async
  error assertion form. Verifies both the error class AND the message.

No behavior change.
