+++ src/utils/validation/resolve-ref.ts
